+++ __tests__/index.ts
@@ -1,2 +1,3 @@
-export * as cloudRun from './cloudrun/index'
+export * as cloudrun from './cloudrun/index'
+export * as cloudrunv2 from './cloudrunv2/index'
 export * as compute from './compute/index'