+++ src/cloudrunv2/index.ts
@@ -1,0 +1,5 @@
+import { cloudrunv2DisallowPublicIngress } from "./cloudrunv2DisallowPublicIngress";
+
+export const cloudrunv2Policies = [
+	cloudrunv2DisallowPublicIngress,
+];